refactor: [YSS-72] 잔소리, 칭찬 대상 조회 클래스 분리
